Naming improvements
